refactor(mc-scripts): extract callback into separate module
