+++ packages/mc-scripts/package.json
@@ -102,7 +102,6 @@
     "postcss-reporter": "7.1.0",
     "prettier": "2.8.8",
     "prompts": "^2.4.2",
-    "qss": "2.0.3",
     "querystring-es3": "^0.2.1",
     "react-dev-utils": "12.0.1",
     "react-refresh": "0.14.2",
