+++ packages/mc-scripts/src/commands/login.ts
@@ -1,18 +1,16 @@
 import crypto from 'node:crypto';
-import http, { IncomingMessage, Server } from 'node:http';
+import type { Server } from 'node:http';
 import chalk from 'chalk';
-import jwtDecode from 'jwt-decode';
 import prompts from 'prompts';
-import { decode } from 'qss';
 import { processConfig } from '@commercetools-frontend/application-config';
+import pkgJson from '../../package.json';
 import { getAuthToken } from '../utils/auth';
+import { createAuthCallbackServer } from '../utils/auth-callback';
 import CredentialsStorage from '../utils/credentials-storage';
-
-type AuthorizeCallbackFragments = { sessionToken?: string; state: string };
-type SessionToken = { exp: number; nonce: string };
 
 const credentialsStorage = new CredentialsStorage();
 const port = 3001;
+const clientId = `__local:${pkgJson.name}@${pkgJson.version}`;
 
 const startServer = (server: Server) =>
   new Promise((resolve, reject) => {
@@ -27,26 +25,6 @@ const startServer = (server: Server) =>
 
 const generateRandomHash = (length: number = 16) =>
   crypto.randomBytes(length).toString('hex');
-
-const parseRequestBody = async (request: IncomingMessage): Promise<string> => {
-  return new Promise((resolve, reject) => {
-    const chunks: Buffer[] = [];
-    request.on('data', (chunk) => {
-      chunks.push(chunk);
-    });
-    request.on('end', () => {
-      try {
-        const body = Buffer.concat(chunks).toString();
-        resolve(body);
-      } catch (error) {
-        reject(error);
-      }
-    });
-    request.on('error', (error) => {
-      reject(error);
-    });
-  });
-};
 
 async function run() {
   const shouldUseExperimentalIdentityAuthFlow =
@@ -72,7 +50,7 @@ async function run() {
     const authUrl = new URL('/login/authorize', mcApiUrl);
     authUrl.searchParams.set('response_type', 'id_token');
     authUrl.searchParams.set('response_mode', 'form_post');
-    authUrl.searchParams.set('client_id', '__local:@@cli@@');
+    authUrl.searchParams.set('client_id', clientId);
     authUrl.searchParams.set(
       'scope',
       [
@@ -83,62 +61,16 @@ async function run() {
     authUrl.searchParams.set('state', state);
     authUrl.searchParams.set('nonce', nonce);
 
-    const server = http.createServer(async (request, response) => {
-      try {
-        if (
-          request.url?.includes('/oidc/callback') &&
-          request.method === 'POST' &&
-          request.headers['content-type'] ===
-            'application/x-www-form-urlencoded'
-        ) {
-          const body = await parseRequestBody(request);
-          const authParams = decode<AuthorizeCallbackFragments>(body);
-          const { sessionToken } = authParams;
+    const server = createAuthCallbackServer({
+      state,
+      nonce,
+      onSuccess: (tokenContext) => {
+        credentialsStorage.setToken(mcApiUrl, tokenContext);
 
-          if (!sessionToken) {
-            throw new Error(
-              'Invalid authentication flow (missing sessionToken)'
-            );
-          }
-          const decodedSessionToken = jwtDecode<SessionToken>(sessionToken);
-
-          if (decodedSessionToken?.nonce !== nonce) {
-            throw new Error('Invalid authentication flow (nonce mismatch)');
-          }
-          if (authParams.state !== state) {
-            throw new Error('Invalid authentication flow (state mismatch)');
-          }
-
-          credentialsStorage.setToken(mcApiUrl, {
-            token: sessionToken,
-            expiresAt: decodedSessionToken.exp,
-          });
-
-          console.log(chalk.green(`Login successful.`));
-          console.log();
-
-          response.writeHead(200, 'Success!', { 'content-type': 'text/html' });
-          response.end();
-          server.close();
-        }
-      } catch (error) {
-        if (error instanceof Error) {
-          console.error(error.message);
-          response.writeHead(500, error.message, {
-            'content-type': 'text/html',
-          });
-        } else {
-          console.error(error);
-          response.writeHead(500, `Something went wrong. Please try again.`, {
-            'content-type': 'text/html',
-          });
-        }
-        response.end();
-
-        server.close();
-      }
+        console.log(chalk.green(`Login successful.`));
+        console.log();
+      },
     });
-
     await startServer(server);
 
     await open.default(authUrl.toString());
